pyproject.toml
[project]
name = "ai-indian-stock-suggestion-vibe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dnspython>=2.8.0",
    "httpx>=0.27.2",
    "email-validator>=2.3.0",
    "fastapi>=0.136.1",
    "openai>=2.14.1",
    "pymongo[srv]>=4.17.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "uvicorn>=0.46.0",
]
